My spaghetti code after first lab. So today I show my code to my teacher. It was very awefull. Nothing was working, despite on that I chek it millions time. So I add to first task method that add student to dormitory. Also I fix a lot of bugs in second task. Now it is working very good. I hope. Plus to this I create divide my map form drawing part and create gui. So now you can play both in console and window. Trully, I fell very bad about my code right now. I have to make it better. I will. Also there are left some tasks to do, so I have three days time. And currently we are righting the documentation for our bot for telegram. It is not even a docs, more like book. Have to finish it untill tommorow. I like mylife. It is very awesome.

+++ RecursiveList/DeleteNode.cs
@@ -14,7 +14,15 @@ namespace OOP_Concepts.RecursiveList
         }
         public void deleteOne(int value)
         {
-            deleteElement(root, value,true);
+            if (root.value == value)
+            {
+                if (root.next != null)
+                    root = new Node(root.next);
+                else
+                    clear();
+            }
+            else
+                deleteElement(root, value,true);
         }
         public void deleteAll(int value)
         {
@@ -45,7 +53,10 @@ namespace OOP_Concepts.RecursiveList
 
             while (n.next.value == value)
             {
-                n.next = new Node(n.next.next);
+                if (n.next.next == null)
+                    n.next = null;
+                else
+                    n.next = new Node(n.next.next);
                 if (isOne)
                     return;
             }
